Complete overhaul of reporter API, finish TAP and Mocha-like spec reporters
It was when I got to designing the reporters that I realized how bad the API
was...

See docs/reporters.md to see the new API. The primary changes are that the
event types have changed a lot in both naming and behavior. Also, for tests
without children, it's simply `pass`, `fail`, or `skip`, with no other events
fired. Passing tests with children are marked via `enter` and `leave` wrapping
their children.

This makes writing reporters much easier to do...

Oh, and this new reporter API is a bit better documented as well.